fix bugs and issues on feedback client 14/06/2022

+++ Models/Modules/Cart/CartPendahuluan.cs
@@ -56,7 +56,7 @@ namespace MSNK.Models.Modules.Cart
         public virtual void AddItem2(
             int SpPendahuluanPelbagaiId,
             int Indek,
-            int Baris,
+            decimal Baris,
             string Perihal,
             decimal Kadar,
             decimal Bil,
